Intento 1 de compilacion de la biblioteca

+++ lib/src/sep_pares.c
@@ -7,12 +7,12 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int *Sep_pares(int **matrizA, int tam){   //function
+int *Sep_pares(int **matrizA, size_t tam){   //function
 
 	int *vector_pares = (int*)malloc(tam*sizeof(int)); 		  //reservamos memoria para el vector_pares
 	int contador = 0;										  //inicializamos un contador para cada posición del vector_pares
-	for(int i = 0; i < tam; i++){							  //recorremos la fila y columnas con los 2 for 
-		for(int j = 0; j < tam; j++){
+	for(size_t i = 0; i < tam; i++){							  //recorremos la fila y columnas con los 2 for 
+		for(size_t j = 0; j < tam; j++){
 			if(matrizA[i][j] %2 == 0 && matrizA[i][j] !=0){	  //verificamos cada posición de la matriz si es un numero par
 				vector_pares[contador] = matrizA[i][j]; 	  //guardamos cada numero par encontrado en la matriz en el vector_pares
 				contador++;   								  //incrementamos en 1 el contador
